Modifying the id type from integer to long.

+++ src/test/java/tran/example/weatherforecast/services/repositoryservices/RoleServiceRepositoryImplTestIT.java
@@ -65,7 +65,7 @@ public class RoleServiceRepositoryImplTestIT {
      */
     @Test(expected = NotFoundException.class)
     public void getByIdFailure() {
-        int invalidId = -1;
+        Long invalidId = -1L;
         roleServiceRepository.getById(invalidId);
     }
